requirements.txt
amqp==5.2.0
asgiref==3.6.0
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==5.2.2
docopt==0.4.0
idna==3.7
importlib-metadata==7.1.0
inflection==0.5.1
kombu==5.3.7
mandrill==1.0.60
Markdown==3.4.3
packaging==24.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==2.6.0
python-dateutil==2.9.0.post0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
zipp==3.18.1
